+++ src/listeners/panel/index.ts
@@ -2,6 +2,7 @@ export * from './ArticleListener';
 export * from './BaseListener';
 export * from './DataListener';
 export * from './ExtensionListener';
+export * from './FieldsListener';
 export * from './MediaListener';
 export * from './ScriptListener';
 export * from './SettingsListener';