+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### ✨ New features
 
 - [#424](https://github.com/estruyf/vscode-front-matter/issues/424): Snippet wrapping to allow easier updates or changes to previously set snippets in the content
+- [#585](https://github.com/estruyf/vscode-front-matter/issues/585): New content relationship field type (`contentRelationship`)
 
 ### 🎨 Enhancements